[EC-73] feat: move access selector to shared
