+++ apps/web/src/app/organizations/organization.module.ts
@@ -4,7 +4,6 @@ import { ApiService as ApiServiceAbstraction } from "@bitwarden/common/abstracti
 
 import { SharedModule } from "../shared";
 
-import { AccessSelectorModule } from "./components/access-selector";
 import { GroupAddEditComponent } from "./manage/group-add-edit.component";
 import { GroupsComponent } from "./manage/groups.component";
 import { UserGroupsComponent } from "./manage/user-groups.component";
@@ -14,12 +13,7 @@ import { GroupServiceAbstraction } from "./services/abstractions/group";
 import { GroupService } from "./services/group/group.service";
 
 @NgModule({
-  imports: [
-    SharedModule,
-    OrganizationsRoutingModule,
-    SharedOrganizationModule,
-    AccessSelectorModule,
-  ],
+  imports: [SharedModule, OrganizationsRoutingModule, SharedOrganizationModule],
   declarations: [
     GroupsComponent,
     GroupAddEditComponent,
