+++ apps/web/src/app/organizations/shared/components/access-selector/access-selector.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from "@angular/core";
 
-import { SharedModule } from "../../../shared/shared.module";
+import { SharedModule } from "../../../../shared";
 
 import { AccessSelectorComponent } from "./access-selector.component";
 import { UserTypePipe } from "./user-type.pipe";
