+++ apps/web/src/app/organizations/shared/components/collection-dialog/collection-dialog.component.ts
@@ -20,7 +20,7 @@ import {
   AccessItemView,
   convertToPermission,
   convertToSelectionView,
-} from "../../../components/access-selector";
+} from "../access-selector";
 
 export interface CollectionEditDialogParams {
   collectionId?: string;
